node_manager/node-agent/requirements.txt
attrs==21.4.0
azure-core==1.23.1
azure-storage-blob==12.11.0
azure-storage-file-share==12.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.1
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.4.0
kafka-python==2.0.2
MarkupSafe==2.1.1
msrest==0.6.21
oauthlib==3.2.0
psutil==5.9.0
pycparser==2.21
pydantic==1.9.0
pymongo==4.1.1
pyrsistent==0.18.1
python-on-whales==0.43.0
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
tqdm==4.64.0
typer==0.4.1
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1